% Auto-generated: do not edit by hand
\name{r_ReactDnd}

\alias{r_ReactDnd}

\title{ReactDnd component}

\description{
ReactDnd takes an array of data and statuses  and displays them in columns for the user to drag and update their status.
}

\usage{
r_ReactDnd(id=NULL, statuses=NULL, data=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{statuses}{Unnamed list. List of status columns for each task}

\item{data}{Unnamed list. A list of current tasks and their status}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

